main-commerce/requirements.txt
alembic==1.13.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
saga_py==1.1.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
pytz==2024.1
tenacity==8.2.3
Flask-Caching==2.1.0
redis==5.0.3
pybreaker==1.2.0